Texture fixes
Removed redundant AASTINKY and AASHITTY texture definitions which overrode the null texture definitions. This caused some linedefs to be visible when it wasn't intended (e.g. FLOOD.WAD).
Added M_UVPLUS graphic to the list of resources to copy, which I somehow missed when importing changes from the texture-fix branch.