Updates to DOOM TITLEPIC related lumps
As the registered DOOM TITLEPIC is without Cyberdemon now, the HELP screens needed to be updated too.

HELP2 became a downgrade to DOOM V1.1 graphics, because there are no letters that overlap the Cyberdemon picture.
The green messages are anyway out of date and HELP2 is obsolete since Ultimate DOOM, as the lump is not used anymore.

The Cacdodemon is removed from the HELP screens as it is with the original graphics.

The id logo is removed from the HELP1 screen as it is with the original graphics.

The TITLEPICs became a change on a cloud in the upper right corner where the original graphic ends. 

Someone noticed that the extended graphics were made on the modified Scuba Steve graphics, that are used in the Unity and Kex release and wanted to change the DOOM letters to the old ones with the red, orange and yellow pixels in it.
While doing this there was a seam at the upper right area. This is corrected now and the middle area is exchangeable again with the older DOOM letter graphics.