Made Master Levels Rejects toggleable
The Master Levels sub-episodes are no longer listed in the new game menu and in the script.
The script will now list bonus maps in the list of episodes.
Made "Read this!" in compatibility options a submenu.

This commit breaks the Master Levels story intermissions. They will need to be reworked as optional.